Implelemt Heap and Overload new/delete.

+++ Kernel/src/IDT/IDT.cpp
@@ -7,7 +7,7 @@ namespace kernel::idt {
 
         // fill IDT-Descriptor
         idt_ptr.limit = (sizeof(IDTEntry) * idt.size()) - 1;
-        idt_ptr.base = reinterpret_cast<uint32_t>(idt.begin());
+        idt_ptr.base = reinterpret_cast<uintptr_t>(idt.begin());
 
         // Clear table
         for (uint32_t i = 0; i < idt.size(); i++) [[likely]]
@@ -16,11 +16,11 @@ namespace kernel::idt {
         // _build_ IDT
         for (const auto& entry : idt_init_table) [[likely]]
             idt[entry.index].set_gate(
-                reinterpret_cast<uint32_t>(entry.handler),
+                reinterpret_cast<uintptr_t>(entry.handler),
                 CODE_SEGMENT_SELECTOR, 
                 FLAGS
             );
 
-        LoadIDT(reinterpret_cast<uint32_t>(&idt_ptr));
+        LoadIDT(reinterpret_cast<uintptr_t>(&idt_ptr));
     }
 } // namespache kernel::idt
